Users Migrations and Seeders

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -34,6 +34,8 @@ class CreateUsersTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('users');
+        if(App::environment() == 'testing'){
+            Schema::dropIfExists('users');
+        }
     }
 }
